Fix linting issues: format code with black and isort

+++ app/services/Serie_temporelle/load_split.py
@@ -1,6 +1,8 @@
 import os
-import pandas as pd
+
 import mlflow
+import pandas as pd
+
 
 def split_data(input_path, output_folder):
     """Charge le dataset global et le split en séries temporelles par cluster."""
@@ -9,19 +11,18 @@ def split_data(input_path, output_folder):
     mlflow.set_experiment("ST-Split")
     with mlflow.start_run(run_name="split_clusters"):
 
-        df = pd.read_csv(input_path, sep=';', parse_dates=['date'], index_col='date')
-        df = df.dropna(subset=['cluster', 'prix_m2_vente'])
-        df['cluster'] = df['cluster'].astype(int)
+        df = pd.read_csv(input_path, sep=";", parse_dates=["date"], index_col="date")
+        df = df.dropna(subset=["cluster", "prix_m2_vente"])
+        df["cluster"] = df["cluster"].astype(int)
 
-        cluster_list = sorted(df['cluster'].unique())
+        cluster_list = sorted(df["cluster"].unique())
         mlflow.log_param("clusters", cluster_list)
 
         for cl in cluster_list:
-            df_cl = df[df['cluster'] == cl].resample('M').prix_m2_vente.mean().dropna()
+            df_cl = df[df["cluster"] == cl].resample("M").prix_m2_vente.mean().dropna()
             out_path = os.path.join(output_folder, f"cluster_{cl}_series.csv")
-            df_cl.to_csv(out_path, sep=';', index=True)
+            df_cl.to_csv(out_path, sep=";", index=True)
             mlflow.log_artifact(out_path)
 
         mlflow.log_metric("n_clusters", len(cluster_list))
         print(f"{len(cluster_list)} séries sauvegardées dans {output_folder}")
-
